pyproject.toml
[project]
name = "openclaw-search-gateway"
version = "0.1.0"
description = "SearXNG++ search gateway for OpenClaw"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "redis[hiredis]>=5.2.0",
    "pydantic-settings>=2.6.0",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.5.0",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12.0",
    "bleach>=6.2.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
browser = ["playwright>=1.49.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.34.0",
    "fakeredis[lua]>=2.26.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gateway"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
